// src/api/models/WebhookSubscription.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type WebhookSubscription = {
    id?: string;
    url?: string;
    events?: Array<string>;
    isActive?: boolean;
    createdAt?: string;
    updatedAt?: string;
};